Finalized the sequential estimation.
Now, the sequential estimation code is working properly. The batch
estimation code is also working properly, however, for some boundary
conditions, the spectral eigenvalues diagonal matrix becomes singular
due to the rapid vanishing of the eigenvalues. This was addressed in the
code by adding a small value to the diagonal matrix when computing its
inverse. However, the performance for this case is still not good. More
investigation on this issue will be performed.